Move SampleUpload logic to its own component
